# test1/docs/swagger.yaml
basePath: /
definitions:
...
  dto.AssignRolePermissionsRequest:
    properties:
      permission_ids:
        items:
          type: integer
        type: array
    required:
    - permission_ids
    type: object
...
  dto.CreateRoleRequest:
    properties:
      code:
        example: admin
        type: string
      description:
        example: 系统管理员
        type: string
      name:
        example: 管理员
        type: string
      status:
        example: 1
        type: integer
    required:
    - code
    - name
    type: object
...
  dto.PermissionResponse:
    properties:
      action:
        example: create
        type: string
      code:
        example: user:create
        type: string
      created_at:
        type: string
      description:
        example: 创建用户权限
        type: string
      id:
        example: 1
        type: integer
      method:
        example: POST
        type: string
      name:
        example: 用户创建
        type: string
      path:
        example: /api/users
        type: string
      resource:
        example: user
        type: string
      status:
        example: 1
        type: integer
      type:
        example: api
        type: string
      updated_at:
        type: string
    type: object
  dto.RoleListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.RoleResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  dto.RoleResponse:
    properties:
      code:
        example: admin
        type: string
      created_at:
        example: "2026-05-09T12:00:00Z"
        type: string
      description:
        example: 系统管理员
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 管理员
        type: string
      status:
        example: 1
        type: integer
      updated_at:
        example: "2026-05-09T12:00:00Z"
        type: string
    type: object
  dto.RoleWithPermissionsResponse:
    properties:
      code:
        example: admin
        type: string
      created_at:
        type: string
      description:
        example: 系统管理员
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 管理员
        type: string
      permissions:
        items:
          $ref: '#/definitions/dto.PermissionResponse'
        type: array
      status:
        example: 1
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.UpdateRoleRequest:
    properties:
      code:
        example: admin
        type: string
      description:
        example: 系统管理员
        type: string
      name:
        example: 管理员
        type: string
      status:
        example: 1
        type: integer
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 基于 Gin 的管理系统 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 管理系统 API
  version: "1.0"
paths:
...
  /api/roles:
    get:
      consumes:
      - application/json
      description: 获取角色列表，支持分页和条件查询
      parameters:
      - description: 角色名称（模糊查询）
        in: query
        name: name
        type: string
      - description: 角色编码（模糊查询）
        in: query
        name: code
        type: string
      - description: 状态（1启用 0禁用）
        in: query
        name: status
        type: integer
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认10）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleListResponse'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取角色列表
      tags:
      - 角色管理
    post:
      consumes:
      - application/json
      description: 创建一个新角色
      parameters:
      - description: 角色信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 创建角色
      tags:
      - 角色管理
  /api/roles/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除角色
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 删除角色
      tags:
      - 角色管理
    get:
      consumes:
      - application/json
      description: 根据ID获取角色详情
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取角色详情
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 根据ID更新角色信息
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 更新角色
      tags:
      - 角色管理
  /api/roles/{id}/permissions:
    get:
      consumes:
      - application/json
      description: 获取指定角色的所有权限
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PermissionResponse'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取角色的权限列表
      tags:
      - 角色管理
    put:
      consumes:
      - application/json
      description: 为指定角色分配多个权限（覆盖原有权限）
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 权限ID列表
        in: body
        name: permissions
        required: true
        schema:
          $ref: '#/definitions/dto.AssignRolePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RoleWithPermissionsResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 角色不存在
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 为角色分配权限
      tags:
      - 角色管理
...
schemes:
- http
swagger: "2.0"